create separate layout component
